the-xray-tracer/python/requirements.txt
-e .
aws-cdk.core==1.74.0
aws-cdk.aws-iam==1.74.0
aws-cdk.aws-lambda==1.74.0
aws-cdk.aws_apigateway==1.74.0
aws-cdk.aws-sns==1.74.0
aws-cdk.aws-sns-subscriptions==1.74.0
aws-cdk.aws-dynamodb==1.74.0
aws-cdk.aws-sqs==1.74.0
aws-cdk.aws-lambda-event-sources==1.74.0
